refactor(server): extract /api/precog/* into a router module
Sixth route-group extraction. 5 handlers (the precog/score scorer +
score/batch/all/accuracy reads), uniform requireAuth, deps = shared pool
+ verifyBrandAccess (both extracted), no local-helper blockers.

Moves them into src/server/routes/precog.js, mounted
`app.use('/api/precog', requireAuth, precogRouter)` (router-level auth).

Routes (byte-identical, guard-verified):
  POST /api/precog/score
  GET  /api/precog/score/:brandProfileId/:contentId
  POST /api/precog/batch
  GET  /api/precog/all/:brandProfileId
  GET  /api/precog/accuracy/:brandProfileId

Handler bodies verbatim; only registration lines changed.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean
- route-inventory guard -> 213 byte-identical (5 paths)
- vitest -> 93/93

Rollback: revert; verbatim move.
